+++ src/main/java/com/zerobase/cafebom/member/service/dto/SigninDto.java
@@ -1,7 +1,7 @@
 package com.zerobase.cafebom.member.service.dto;
 
 import com.zerobase.cafebom.member.security.Role;
-import com.zerobase.cafebom.member.controller.dto.SigninForm;
+import com.zerobase.cafebom.member.controller.form.SigninForm;
 import lombok.Builder;
 import lombok.Getter;
